// src/app/models/permissions.model.ts
export interface UserPermissions {
  canCreateGroup: boolean;
  canEditGroup: boolean;
  canDeleteGroup: boolean;
  canAddMembers: boolean;
  canRemoveMembers: boolean;
  canCreateTickets: boolean;
  canEditTickets: boolean;
  canViewTicketDetail: boolean;
  canManageGroups: boolean;
  canManageUsers: boolean;

}